requirements.txt
APScheduler==3.3.0
argcomplete==1.7.0
base58==0.2.4
beautifulsoup4==4.5.1
boto3==1.4.2
botocore==1.4.87
bs4==0.0.1
click==6.6
docutils==0.12
Flask==0.12
funcsigs==1.0.2
futures==3.0.5
hjson==2.0.2
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
kappa==0.6.0
lambda-packages==0.11.0
MarkupSafe==0.23
nose==1.3.7
oauthlib==2.0.1
placebo==0.8.1
psycopg2==2.6.2
pyaml==15.8.2
python-dateutil==2.6.0
python-lambda==0.5.0
python-slugify==1.2.1
pytz==2016.10
PyYAML==3.12
requests==2.12.4
requests-oauthlib==0.7.0
s3transfer==0.1.10
six==1.10.0
toml==0.9.2
tqdm==4.10.0
troposphere==1.9.0
tweepy==3.5.0
tzlocal==1.3
Unidecode==0.4.19
Werkzeug==0.11.11
wsgi-request-logger==0.4.6
zappa==0.32.1
